assign unique id to each logical plan node

+++ xngin-protocol/src/mysql/serde/mod.rs
@@ -191,7 +191,7 @@ impl From<u32> for LenEncInt {
         } else if src <= 0xffff {
             LenEncInt::Len3(src as u16)
         } else {
-            LenEncInt::Len4(src as u32)
+            LenEncInt::Len4(src)
         }
     }
 }
